Added 'updateHeightMap()' to 'GlobeSphere'
The 'updateHeightMap' method adjusts the radius of each point from the center according to the height map image provided. Now, both the Cube and Globe Spheres have this capability.

In addition, PImage_Help has the ability to load the pixels in an image to a 'pixels2D' array. This allows for easier referencing specific pixels instead of calculating it.