Code cleanup and added code for cube to globe conversion
Code was cleaned up and comments were added.

'triangleInterp' finds the equation for the 3D plane that intersects 3 vertices (with the value being the z coord) in a triangle and then calculates the value at a specific location (pixel location).

'isInsideTriangle' just determines if a point [x,y] is in a triangle defined by points p1, p2, and p3.